+++ changelog.md
@@ -1,6 +1,9 @@
-2023-07-13 1.4.4
+2023-07-16 1.4.4
  - Minimizes reported click/pops on exported chains when importing wav files of differing sample rate to the target sample rate, or when using audio interfaces configured with small sample buffers. On by default at a 40% threshold, user configurable thresholds and can be disabled via settings panel.
  - Resolves issue where running in the Brave browser caused an unexpected error modifying an existing audio array buffer.
+ - Added fuzz selected actions action to add fuzz/noise into the selected samples.
+ - Added fade-in fade-out selected actions to fade in/out the first/last 256 samples of each selected sample.
+ - Blend panel to blend/interpolate between the selected samples (works best when the samples are similar in duration).
 
 2023-07-02 1.4.3
  - Added a ping-pong selected actions action to add a reversed copy of the sample to the end of the sample.